Switch to Port >1024 on Unprivileged Containers
There can be configurations where running unprivileged
containers may result in containers not starting up
properly. This has mainly to do with ports <=1024 not
being allowed to be bound by unprivileged users.
By choosing a port >1024 one does not need to have root
privileges in order to bind it.

This change ensures that this project can be ran on a
broader set of machine configurations.